requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
